when there are error entries in dns_log, status lit to red.
when user fixed those errors, subsequent dns_log entries were green, but
the code picked up on reds from the past anyway.

make the visual status bipolar, first part is for the very last log line,
another for the highest problematic of the rest of the log lines.

this should be improved to something better, perhaps counting number of
entries of each kind and colouring based on percentage.